+++ truffle/src/com.oracle.truffle.api.vm/src/com/oracle/truffle/api/vm/PolyglotLanguageContext.java
@@ -34,7 +34,6 @@ import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.function.BiFunction;
-import java.util.function.Function;
 
 import org.graalvm.polyglot.Value;
 import org.graalvm.polyglot.impl.AbstractPolyglotImpl.APIAccess;
@@ -57,7 +56,7 @@ final class PolyglotLanguageContext implements PolyglotImpl.VMObject {
 
     final PolyglotContextImpl context;
     final PolyglotLanguage language;
-    volatile Map<Source, CallTarget> sourceCache;
+    volatile PolyglotSourceCache sourceCache;
     final Map<String, Object> config;
     final boolean eventsEnabled;
     volatile Map<Class<?>, PolyglotValue> valueCache;
@@ -96,11 +95,11 @@ final class PolyglotLanguageContext implements PolyglotImpl.VMObject {
         defaultValueCache = new PolyglotValue.Default(this);
 
         assert language.isInitialized();
-        Map<Source, CallTarget> languageSourceCache = language.sourceCache;
+        PolyglotSourceCache languageSourceCache = language.sourceCache;
         if (languageSourceCache != null) {
             this.sourceCache = languageSourceCache;
         } else {
-            this.sourceCache = new ConcurrentHashMap<>();
+            this.sourceCache = new PolyglotSourceCache();
         }
     }
 
@@ -152,29 +151,11 @@ final class PolyglotLanguageContext implements PolyglotImpl.VMObject {
         return env != null && initialized;
     }
 
-    CallTarget parseCached(PolyglotLanguage accessingLanguage, com.oracle.truffle.api.source.Source source, String[] argumentNames) throws AssertionError {
+    CallTarget parseCached(PolyglotLanguage accessingLanguage, Source source, String[] argumentNames) throws AssertionError {
         ensureInitialized(accessingLanguage);
-        assert this.sourceCache != null;
-
-        if (argumentNames == null || argumentNames.length == 0) {
-            return this.sourceCache.computeIfAbsent(source, new Function<Source, CallTarget>() {
-                public CallTarget apply(Source t) {
-                    return parseImpl(t, null);
-                }
-
-            });
-        } else {
-            // cache is not implemented for argument names
-            return parseImpl(source, argumentNames);
-        }
-    }
-
-    private CallTarget parseImpl(com.oracle.truffle.api.source.Source t, String[] argumentNames) throws AssertionError {
-        CallTarget target = LANGUAGE.parse(requireEnv(), t, null, argumentNames);
-        if (target == null) {
-            throw new AssertionError(String.format("Parsing resulted in a null CallTarget for %s.", t));
-        }
-        return target;
+        PolyglotSourceCache cache = this.sourceCache;
+        assert cache != null;
+        return cache.parseCached(this, source, argumentNames);
     }
 
     Env requireEnv() {