+++ sdk/CHANGELOG.md
@@ -7,6 +7,7 @@ This changelog summarizes major changes between Graal SDK versions. The main foc
 * Added `Context.getCurrent()` to lookup the current context to allow Java methods called by a Graal guest language to evaluate additional code in the current context.
 * Removed deprecated `Context.exportSymbol` and `Context.importSymbol`.
 * Removed deprecated `Source.getCode`.
+* The code cache for sources is now weak. Code can be garbage collected if a source is no longer referenced but the Context or Engine is still active.
 
 ## Version 1.0 RC1
 * Added Context.Builder#allowHostClassLoading to allow loading of new classes by the guest language.